Refactor analytics-context into its own folder and move events into its own folder
