+++ src/contexts/analytics-context/events/misc-events.js
@@ -1,0 +1,7 @@
+export function logout() {
+    this.analytics.trackEvent({
+        category: "UI Interaction",
+        action: "Logout",
+        label: `User logged out.`
+    });
+}
