+++ src/contexts/analytics-context/events/routing-events.js
@@ -1,0 +1,9 @@
+export function trackLocation(location) {
+    this.analytics.trackRoute({
+        route: location.pathname,
+        customParameters: {
+            "URL origin": location.origin,
+            "URL search params": location.search
+        }
+    });
+}
